Add quiz what is missing game
Concept is that you are quizzed on every item from a particular set in
random order but leaving out one item. At the end you must enter what was
missing.

The way you do this is by already having a memory palace of the items and
when you are read each item, you visually imagine the item on fire or some
other sort of mangling as a way of ticking it off. At the end, you review
the memory palace and identify which one was never mangled.

Initial data set is hard-coded to quiz from the books of the Bible. Can
later extend it to signal which input to use.